feat(pci): enhance MSI-X handling with new table entry count and offset methods

+++ arch_hal/aarch64_hal/soc/src/bcm2712/mod.rs
@@ -130,11 +130,9 @@ fn search_rp1(view: &DtbNodeView) -> WalkResult<Rp1Config, Bcm2712Error> {
                 }
 
                 let msi_x = unsafe { brcm_stb.get_msi_x_capability() }?;
-                let msi_x_bar = unsafe {
-                    brcm_stb.msi_x_table_bar_addr(msi_x)?
-                } as u64;
-                let bar1 = unsafe { brcm_stb.read_bar_address(1) }? as u64;
-                let bar2 = unsafe { brcm_stb.read_bar_address(2) }? as u64;
+                let msi_x_bar = unsafe { brcm_stb.msi_x_table_bar_addr(msi_x)? };
+                let bar1 = unsafe { brcm_stb.read_bar_address(1) }?;
+                let bar2 = unsafe { brcm_stb.read_bar_address(2) }?;
                 println!("PCIE: RP1 Pheripheral BAR addr: 0x{:x}", bar1);
                 println!("PCIE: RP1 Shared SRAM BAR addr: 0x{:x}", bar2);
